+++ cjkcms/static/cjkcms/js/cjkcms-front.js
@@ -3,71 +3,63 @@ CJK CMS based on CodeRed CMS
 @license magnet:?xt=urn:btih:c80d50af7d3db9be66a4d0a86db0286e4fd33292&dn=bsd-3-clause.txt BSD-3-Clause
 */
 
-// libs = {
-//     modernizr: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/modernizr/2.8.3/modernizr.min.js",
-//         integrity: "sha256-0rguYS0qgS6L4qVzANq4kjxPLtvnp5nn2nB5G1lWRv4=",
-//     },
-//     moment: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.24.0/moment.min.js",
-//         integrity: "sha256-4iQZ6BVL4qNKlQ27TExEhBN1HFPvAvAMbFavKKosSWQ="
-//     },
-//     pickerbase: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/picker.js",
-//         integrity: "sha256-hjN7Qqm7pjV+lms0uyeJBro1vyCH2azVGqyuWeZ6CFM=",
-//         head: '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/themes/default.css" integrity="sha256-wtVxHQXXtr975G710f51YDv94+6f6cuK49PcANcKccY=" crossorigin="anonymous" />'
-//     },
-//     pickadate: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/picker.date.js",
-//         integrity: "sha256-Z4OXXhjTbpFlc4Z6HqgVtVaz7Nt/3ptUKBOhxIze1eE=",
-//         head: '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/themes/default.date.css" integrity="sha256-U24A2dULD5s+Dl/tKvi5zAe+CAMKBFUaHUtLN8lRnKE=" crossorigin="anonymous" />'
-//     },
-//     pickatime: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/picker.time.js",
-//         integrity: "sha256-mvFcf2wocDC8U1GJdTVSmMHBn/dBLNeJjYRvBhM6gc8=",
-//         head: '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/pickadate.js/3.6.3/compressed/themes/default.time.css" integrity="sha256-dtpQarv++ugnrcY7o6Gr3m7fIJFJDSx8v76jjTqEeKE=" crossorigin="anonymous" />'
-//     },
-//     jquery_ui: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.js",
-//         integrity: "sha256-T0Vest3yCU7pafRw9r+settMBX6JkKN06dqBnpQ8d30=",
-//     },
-//     jquery_qtip: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/qtip2/3.0.3/basic/jquery.qtip.min.js",
-//         integrity: "sha256-219NoyU6iEtgMGleoW1ttROUEs/sux5DplKJJQefDwE=",
-//     },
-//     fullcalendar: {
-//         url: "https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.10.0/fullcalendar.min.js",
-//         integrity: "sha256-4+rW6N5lf9nslJC6ut/ob7fCY2Y+VZj2Pw/2KdmQjR0=",
-//         head: '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.10.0/fullcalendar.min.css" integrity="sha256-9VgA72/TnFndEp685+regIGSD6voLveO2iDuWhqTY3g=" crossorigin="anonymous" />' +
-//               '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/3.10.0/fullcalendar.print.min.css" media="print" integrity="sha256-JYJWCNB1pXBwUngem7hITwB6SdmCGkhewhKS8NL1A8A=" crossorigin="anonymous" />'
-//     },
-//     coderedmaps: {
-//         url: "/static/coderedcms/js/codered-maps.js",
-//         integrity: "",
-//     },
-//     coderedstreamforms: {
-//         url: "/static/coderedcms/js/codered-streamforms.js",
-//         integrity: "",
-//     }
-// }
-
-// function load_script(lib, success) {
-//     // lib is an entry in `libs` above.
-//     // It is best to put functionality related to the script you are loading into the success callback of the load_script function.
-//     // Otherwise, it might not work as intended.
-//     if(lib.head) {
-//         $('head').append(lib.head);
-//     }
-//     if(lib.url){
-//         $.ajax({
-//             url: lib.url,
-//             dataType: "script",
-//             integrity: lib.integrity,
-//             crossorigin: "anonymous",
-//             success: success
-//         });
-//     }
-// }
+/**
+ * Main script which is used to detect Cms features requiring JavaScript.
+ *
+ * Loads the necessary libraries for that feature, then initializes any
+ * feature-specific code. This should only be used for features that might be
+ * site-wide (e.g. StreamField blocks that could occur anywhere). For
+ * functionality that is page-specific, include the JavaScript normally via a
+ * script tag on that page instead.
+ *
+ * This file must run with "pure" JavaScript - assume jQuery or any other
+ * scripts are not yet loaded.
+ */
+const libs = {
+    masonry: {
+      url: "https://cdn.jsdelivr.net/npm/masonry-layout@4.2.2/dist/masonry.pkgd.min.js",
+      integrity: "sha256-Nn1q/fx0H7SNLZMQ5Hw5JLaTRZp0yILA/FRexe19VdI=",
+    },
+  };
+  
+  /**
+   * Dynamically loads a script and/or CSS from the `lib` object above.
+   *
+   * Put functionality related to the script you are loading into the `success`
+   * callback of the `load_script` function. Otherwise, it might not work as
+   * intended.
+   */
+  function load_script(lib, success) {
+    var head = document.getElementsByTagName("head")[0];
+    if (lib.head) {
+      // Create a temporary element and insert the `lib.head` string to form a
+      // child element.
+      var tmpEl = document.createElement("div");
+      tmpEl.innerHTML = lib.head;
+      // Append the child element to the `<head>`
+      head.append(tmpEl.firstElementChild);
+    }
+    if (lib.url) {
+      // Fetch and execute the script in the global context.
+      // Then call the `success` callback.
+      fetch(lib.url, {
+        integrity: lib.integrity,
+        referrerPolicy: "origin",
+      })
+        .then(function (response) {
+          return response.text();
+        })
+        .then(function (txt) {
+          // Eval in the global scope.
+          eval?.(txt);
+        })
+        .then(function () {
+          if (success) {
+            success();
+          }
+        });
+    }
+  }
 
 document.addEventListener("DOMContentLoaded", function() {
     /*** Link handling ***/
@@ -103,7 +95,7 @@ document.addEventListener("DOMContentLoaded", function() {
 
     /** Tracking **/
     
-    if (typeof cr_track_clicks !== "undefined" && cr_track_clicks) {
+    if (typeof cms_track_clicks !== "undefined" && cms_track_clicks) {
         document.querySelectorAll("a").forEach(function (el) {
             el.addEventListener("click", function (event) {
             var el = event.currentTarget;
@@ -121,5 +113,43 @@ document.addEventListener("DOMContentLoaded", function() {
             });
         });
     }
+
+    /** Masonry **/
+    if (document.querySelectorAll("[data-masonry]").length > 0) {
+        load_script(libs.masonry);
+    }
+
+    /** Film Strip Controls **/
+    let strips = document.querySelectorAll("[data-block='film-strip']");
+    strips.forEach((el) => {
+        const leftButton = el.querySelector("[data-button='left']");
+        const rightButton = el.querySelector("[data-button='right']");
+        const container = el.querySelector("[data-block='film-container']");
+
+        leftButton.addEventListener("click", function () {
+            const panels = el.querySelectorAll("[data-block='film-panel']");
+            let currentBlock = parseInt(el.dataset.currentBlock) - 1;
+            if (currentBlock < 0) currentBlock = panels.length - 1;
+            el.dataset.currentBlock = currentBlock;
+
+            const elem = panels[currentBlock];
+            const left = elem.offsetLeft;
+
+            container.scroll({ top: 0, left: left, behavior: "smooth" });
+        });
+
+        rightButton.addEventListener("click", function () {
+            const panels = el.querySelectorAll("[data-block='film-panel']");
+            let currentBlock = parseInt(el.dataset.currentBlock) + 1;
+            if (currentBlock >= panels.length) currentBlock = 0;
+            el.dataset.currentBlock = currentBlock;
+
+            const elem = panels[currentBlock];
+            const left = elem.offsetLeft;
+
+            container.scroll({ top: 0, left: left, behavior: "smooth" });
+        });
+    });
+
 });